testing/web-platform/tests/old-tests/submission/Infraware/Forms/contents/Forms/keygen_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>keygen_form</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form onsubmit="return false" name='input_form'>
   <p><keygen name="key" id='keygen_id'></keygen></p>
  </form>

  <script>
    
    var keygen = document.getElementById("keygen_id");
    var form = keygen.form;

    if (typeof(form) == "object") {
      test(function() {
        assert_equals(form.name, "input_form", "form attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("form attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
